DDPB-2538: Confirm report unsubmission (#1187)
* Add confirmation page template

* Defer to confirmation page once form first submitted

* Add translations for new text content

* Add edit links to go back

* Fix calculation of submitted due date

* Add confirmation field to form

* Add 'reportId' to template

This is necessary because the report is overwritten by `$form->handleRequest()`, causing the ID to be unset. A better solution might be to stop tying the form to `$report` as tightly as it currently is.

* Create separate confirmation form

* Breakdown confirmation stages of manageAction better

* Simplify form rendering

* Store unsubmit dates as strings

In order to use `HiddenType` for fields so they don't render on the page, we need to store start, end and due dates as strings. This means converting them from dates when populating the confirmation page and back _to_ dates when actually performing the save.

* Replace `->get` with direct array access

* Directly reference form rather than using unnecessary holder

* Update tests to support new confirmation screen

* Update warning message

* Retain data when going back with 'Edit' link

* Handle cases where custom date not set

* Use section IDs rather than index