DDPB-2622: Alter email tests to no longer clear log (#1218)
* Remove unused email test traits

* Update 'no email should have been sent' tests

Now specifically checks what happened to last email, which means the email log doesn't need to be reset

* Update email test traits to reflect internal changes

* Fix whitespace in feature file

* Remove invalid test steps

The test implies it checks that deputy users cannot use the admin password reset form. This is not true, either form works and the user will be sent a reset email to their allocated area.

The test passes because it's been configured wrong. It actually does:
- Set area to deputy
- Reset emails
- Reset password from admin form
- Check no email was sent

This passes because no email was sent to the deputy area: it was sent to the admin area instead. To properly configure the test, it should set the area to admin before attempting reset, but that would fail because there is no restriction in the code as mentioned above.

* Remove 'I reset the email log' test steps

* Remove 'I reset the email log' test trait

* Only iterate over emails if there were any