DDPB-2556: Use common base docker images rather than prebuilt ones (#548)
* Split docker-compose file from Docker repo

* Add `log` to .gitignore

* Copy environment variables into repo

* Use local environment variables rather than external repo

* Add composer container to docker-compose

* Update generated files with new composer build

We're using a newer version of composer, so the built files are slightly different

* Add new nginx container for API

Previously Nginx was installed on the PHP image, we've now separated them. Also implements a new 'digideps' network which allows frontend and API to communicate.

* Use global base images rather than internal registry

We now have access to the entire build chain for the API, rather than using an image from our internal blackbox registry.

* Remove unused conf files

* Output monolog to stderr rather than a file

* Update setup instructions

* Move cache/logs to `var` directory

`var` isn't mounted in Docker, so all the I/O needed for cache and logs doesn't need to be synced by Docker. This leads to _extraordinary_ performance increases.

* Add Redis client to API build

* Remove old logs directory

Logs are now stored in `var`, which is not mounted

* Ensure correct directories exist on container

* Add instructions for installing dependencies to README

* Stop tracking cache and log files now in `var`

* Reintroduce NGINX environment variables

* Update script file locations and users

Also ensure scripts are available in Docker

* Make tests available to API container

* Add postgresql-client to PHP container

This allows it to run commands like `psql` and `psql_dump`

* Output test logs to a file

* Downgrade Postgres to 9.5.13

Alpine client supports 9.5.13 client, so we need to match

* Better parameterise NGINX config

* Remove requirement for API_FIXTURES to be a dir

Effectively means you can continue with just one env variable prefixed API_FIXTURES_ rather than needing a whole array of them

* Remove unused environment variables

* Directly reset database rather than using my_init.d

* Increase memory limit for PHPUnit tests

I tried to avoid doing this, because PHP memory leaks are normally an indication of something else going wrong. But in this case, it's actually matching the configuration currently in use.

* More clearly define cache location

* Remove deprecated links parameters

Everything on the same network can talk to each other, specific links are no longer necessary.

* Separate virtual host and general nginx config files

* Move composer, NPM and nginx into PHP images

* Remove unused web and composer docker config

* Separate Composer into multi-stage build

* Copy entire Dockerfile onto container

* Update script directories and users

* Update missing script user

* Use `confd` to configure environment in scripts

Used to use my_init.d, but that's removed now

* Add default value for OPG_NGINX_INDEX

As well as providing a bit of protection against misconfiguration, this allows hosts to run PHP scripts without needing the full NGINX context

* Add unit tests to API image

* Setup parameters.yml before running API unit tests

* Generate optimised files from composer

* Enable Opcache and pool variables

* Push nginx logs to stdout/err

* Tidy up docker-compose file

* Remove `my_init.d` files

We run scripts more explicitly now

* Update setup instructions in README

* Don't automatically load fixtures

Otherwise we would end up trying to load them on production 🙊

Also, create cache and log directories as part of build

* Upgrade API PHP to 5.6

Minor version upgrade to match frontend

* Update opcache reference

When I upgraded to PHP5.6, the directory containing the opcache extension moved. I've now updated it to point to the correct directory.

* Remove unused docker-compose.yml file

* Ensure PHP-FPM output included in logs

* Install su-exec on API image

This allows us to run commands as other users. It's essential, because otherwise things like `php app/console ...` are run as root, in which case all the directories are created with root permissions. This caused an issue with caches not being accessible by the application.

* Run Symfony console commands as www-data

This ensures that cache files and directories are created under the correct user