DDPB-2648: Additional option for deputy declaration (#1232)
* Add new declaration terms to report submission

* Add new declaration terms to NDR submission

* Add new declaration terms to formatted reports

Also, fix a bug with non-NDR reports which were previously using the wrong translation key

* Add new declaration option

* Upgrade Symfony to 3.4.28 and Twig to 1.41.0

To avoid security vulnerabilities present in previous versions

* Uninstall unused gulp-watch dependency

* Fix issue with Gulp watch

We have to use polling because the watch is inside a Docker container, not on the native machine

* Don't try and delete access token if infeasible

Somewhere between Symfony 3.4.11 and 3.4.28, there was a change that meant the order of logout listener being called may be different. This means that the user token may be an `AnonymousToken` by the time the logout listener is called. Therefore, we should check that the token is a `UsernamePasswordToken` before trying to invalidate it against the API.

We don't _need_ to manually invalidate the login token at logout because they automatically expire after an hour (as would happen if the user closed the tab rather than logging out).

* Update tests referring to declaration options

As we've inserted a new option at the top of the list, all the IDs have incremented and need to be pushed up

* Revert "Add new declaration terms to NDR submission"

NDR users should not see the new declaration option, it's just for Prof and PA users.

This reverts commit 8481e161760167742f272f92d2ca2e09928a9600.

* Revert changes to NDR test

New declaration option should not be shown to NDR users

* Use new declaration text only for non-lay deputies

* Depend on report type, not user type

* Remove 'not_deputy' option from NDR declaration

* Only show 'not_deputy' option for non-lay reports

* Add `isLayReport` property to reports

* Fix capitalisation on translation strings

* Add appropriate classes to formatted report

* Fix `isLayReport` property

* Remove unneeded NDR translations

* Fix tests with new logic

Had previously been changed under logic that lay deputies would also see these options

* Add unit tests for `isLayReport`

* Port `isLayReport` to API

* Show declaration text based on submitting user, not report type

* Show additional declaration option based on report type, not user

* Deprecate `isLayReport`