DDPB-2646: Install GOV.UK Design System as future upgrade for GOV.UK Elements (#1241)
* Install `govuk-frontend` dependency

* Add GOVUK frontend CSS

* Add GOVUK frontend JavaScript and assets

* Initialise GOVUK frontend JavaScript

* Allow govuk-elements and govuk-frontend work side-by-side

We're not going to be able to make a straight swap, so need to ensure they can both be used at once.

This overwrites a rule in govuk-elements which was setting the `html` font-size to small. This conflicted with govuk-frontend's use of rems. The govuk-elements rule was just for IE6/7 compatibility, which is no longer an issue.

* Set GovUK assets path to relative

We build our assets in a timestamped directory, so need to use a relative path. This ensures fonts and images will be loaded properly.

* Always produce sourcemaps

* Add CSS collector design

* Scrape CSS class data from response

* Show class usage counts in profiler

* Tidy up colouring of CSS class toolbar unit

* Convert header and navigation to govuk-frontend

Mostly just swapping classes around, also tidying up syntax and removing unnecessary elements

* Convert phase banner to govuk-frontend

* Convert skip link to govuk-frontend

Also convert cookie banner to `app-` class name

* Add recording of `app-` CSS classes in profiler

* Hide cookie banner when JS enabled

* Convert layout to govuk-framework

* Convert footer to govuk-frontend

* Remove some deprecated classes

* Prefix `helpline` class

* Convert related items to an `app-` class

* Don't include behat classes in profiler totals

* Add CSS policy document

* Update layout classes to use GOV.UK Design System

* Support new icon names in CSS

And remove unused icons

* Add support for new alert CSS classes

* Revert "Support new icon names in CSS"

This reverts commit 8e1bdc6e0afea249f646df894891704d0e78e7f6.

* Support new icon names in CSS

Also, remove unused icon

* Switch Alert macro to use new CSS classes

* Convert _button and _input components to GOV.UK Design System

* Update login page to use GOV.UK Design System

* Remove unnecessary `js-popups` class

* Convert form error components to GOV.UK Design System

* Link hints, errors and inputs with ARIA

* Convert _select component to GOV.UK Design System

* Don't add any error markup if there are no errors

* Use `govuk-input--width-4` instead of custom `account-number`

* Convert _known-date and _sort-code components to GOV.UK Design System

* Use <legend> on stats page rather than header

* Add error class to sort code inputs

* Convert checkboxes to use GOVUK Design System

* Add new preferable conditional radios syntax

* Render multiline text input as <textarea>

* Clean up timeout popup CSS and JS

* Simplify `form_checkbox` template

Also, allow for both radios and checkboxes (they have different classes in the GOV.UK Design System) and fix a regression with sm
all checkboxes

* Tidy up "administrator access" field for org admins

It was rendering as a radio button under the new GOV.UK framework, so I changed it from a ChoiceType (implied Radio) to CheckboxType (implied Checkbox)

* Standardise CSS for page title and captions

* Convert `_link-button` component to GOV.UK Design System

* Stop forcing `button-link` class on `_link-button` uses

* Switch breadcrumb macros to GOV.UK Design System

* Update back link macro to GOV.UK Design System

* Update test steps to account for DOM changes

* Update tests to account for DOM changes

* Note in CSS Policy that reusable templates and macros should be used

* Update package-lock.json

Now generated against production Node, so more reliable

* Add a specific note about styling conditional radios

* Remove unused variable

* Use SCSS Ampersand to reduce duplication

* Switch from `app-` prefix to `opg-`

* Commit to component-based SCSS files

* Stop treating error messages as raw

* Wrap error summary in `govuk-!-width-two-thirds`

This ensures it doesn't collide with the helpline box, emulating some custom CSS from before the GOV.UK Design System upgrade

* Point header link to https://www.gov.uk/

* Bump `govuk-frontend` to 2.13.0

* Use sass-lint format

The old code was using scss-lint format, despite not using that dependency any more. Since `linters` isn't the config parameter any more, this meant all the config was ignored and the default was used instead.

* Convert SCSS lint rules into SASS lint rules

Generally:
 - If _our_ rule was more specific, I converted it.
 - If the _default_ rule was more specific, I removed it.

This means that we're closer aligned with the default sass-lint rules, and also have very specific rules.

* Allow BEM formatted CSS class names

* Override header logo/content distribution on wider headers

25% of 1200px is approximately equal to 33% of 960px, so the amount of space reserved for the logo is comparable

This gives more space for the links/navigation

* Better white space handling of summary tables

Long-word cells will now compress and break, with hyphens where supported. The "Edit" and/or "Remove" buttons will never be broken.