DDPB-2629: Upgrade PHP to 7.3 (#1253)
* Remove zend_extension commands

These are handled by `docker-php-ext-install`

* Upgrade Docker build images to PHP 7

* Remove Composer PHP5 restriction

This allows a lot of dependencies to bump-update to versions which didn't support PHP5

* Ensure `count()` deals with arrays

PHP7 can't count null

* Upgrade JMS Serializer to 3.x

Deals with a latent performance issue from version 2.x.

NB: Also had to specify serialization name for `does_client_have_a_care_plan`, because 3.x expects it to be `does_client_have_acare_plan` (due to the way it handles capitals)

* Ensure all twig files using macros import them

As of Twig 2.x, imports are handled at a block level so macros must be specifically imported into every file that uses them.

* Use new Twig `replace` syntax

* Fix LinksTrait test syntax

PHP5 would allow you to `count()` anything, in PHP7 it has to be an array

* Specify serialized name for NDR `doesClientHaveACarePlan` property

This was already done for reports in cf6a7d1ec

* Ensure emails exist before trying to count them

PHP5 allowed you to `count()` null, but that throws an error in PHP7

* Replace `catch (Exception)` with `catch (Throwable)`

PHP7 split errors and exceptions. This ensures that both are caught, as before.

* Ensure team names are not null before trying to count them

We previously got away with this because PHP5 treated `count(null)` as 0

* Exclude `cashAssetTitles` from serialization

As a static property, there's no point serializing it

* Reintroduce specification of PHP version in composer

* Initialise `array` type entity properties as arrays

Otherwise they're deserialised as an array if there's daat and `null` otherwise. Using the same type makes our code more consistent and means we don't run into issues trying to `count()` what's there.