DDPB-2797: Complete 103 and 104 reports rather than fudging data (#1293)
* Remove duplicate section

* Fix comment typo

* Add `deputy-103` test tag

* Start 103 report from scratch

Rather than appropriating 102 report

* Fill in already tested report sections

These sections were tested in the 102 report. We just need to fill them in with minimum viable information to test whether the report is submittable.

* Add `deputy-104` test tag

* Fill in 104 report in full

Rather than appropriating the 102 report, we just do a minimum viable 104 report

* Assert 104 client details in 104 tests

We're using the 104 client now, not reusing the 102

* Remove 102 assertions

These assertions made sense when we appropriated the 102 report, but now the data isn't set up like this

* Remove unused test features

* Remove unnecessary report type changes

The reports are already the specified type, so these lines do nothing. With their removal the test step definition can be dropped too since it's not being used anywhere else.