Don't buffer requests in Nginx
With this setting enabled, Nginx would buffer the response to a temporary file then pass it on to PHP. This adds unnecessary work for Nginx, particularly for large files which PHP would reject anyway.

By turning the setting off it's up to PHP to reject files which are too large, which it can do in a more user-friendly way.