DDPB-2978: Add Review Checklist "final decision" question (#112)
* Add support for a fullReview field on checklist

* Add FullReviewChecklist model

* Add finalReviewDecision field

* Serialise FullReviewChecklist

* Add types for full review

One is the overall type, the other is a subtype for handling the review questions

* Fix JMS group of fullReviewDecision

* Add basic full review checklist form

* Support saving of full review checklist

* Use new "opg-sticket-menu" class

This uses the "sticky" CSS position, which works better than our hacky method of applying a class. However, we'll have to add a polyfill for older browsers

* Move sticky-menu class into CSS

* Add polyfill for position:sticky

* Update checklist sidebar to match prototype

Handle classes more consistently, simplify `.opg-related-items`, update order and headings

* Add sidebar for full review checklist

* Fix imports

* Only validate full review checklist on submit

You can save an in progress checklist without needing to pass validation

* Move review checklist into new entity

* Add API endpoints for upserting review checklists

* Update frontend to use new ReviewChecklist entity

* Fix validation of ReviewChecklist

* Remove getters/setters from checklist model

There is very little risk in using public properties for this. The one thing it lacks is strong type control, but this is enforced on validation by @Assert and on saving by @JMS\Type

* Ensure review sidebar has last-modified info

* Separate out lodging and review heading translations

* Unify sidebars in one template

Most of their content is duplicated, so create a template and just vary the bits that change

* Remove now unnecessary if statement

We always redirect/refresh the page so won't get double-notice issues

* Remove unnecessary fallback

The API call will _always_ return a review checklist; even a blank one

* Save submission information for review checklist

* Distinguish checklists in automated tests

* Update button label to match story AC

* Reinstate save buttons after submission

It's not part of the AC to remove them and it confuses the application much more, especially at this stage

* Fix typo with migration name

* Specify region for "Save progress" button

Because there are two on the page, one for each checklist

* Specify correct buttons in full review tests

* Add new checklist-accessing test step

Add "I open the :period checklist for client :caseNumber" to navigate to a client's checklist without unnecessary verbosity

* Combine duplicate route commands

* Rename type to match entity underneath

* Specify correct button for lodging checklist tests

* Use helper test set in older tests