Ignore grep exit code
If grep doesn't find anything, it returns an exit code of 1. This causes the command to be logged as a failure, but not finding any PHP changes is an acceptable situation.

Instead, we now continue as normal if the exit code ($?) was 1. If any other exit code occurs, we assume that's a real error and fail as usual.