Use PHPStan to analyse source code for problems (#128)
* Install phpstan

* Check phpstan in CI

Only check updated files, since there are thousands of errors

* Ignore grep exit code

If grep doesn't find anything, it returns an exit code of 1. This causes the command to be logged as a failure, but not finding any PHP changes is an acceptable situation.

Instead, we now continue as normal if the exit code ($?) was 1. If any other exit code occurs, we assume that's a real error and fail as usual.

* Add instructions for using PHPStan locally

* Change file to trigger PHPStan in CI

* Add debug steps to CI

* Update debug

* Remove debug steps

* Specify entrypoint for PHPStan

* Debug what's in bin directory

* Don't copy `bin` from local filesystem

It's automatically generated by composer

* Copy bin files from composer

* Remove debugging steps

* Remove bin files

They should be generated by composer, not hard-coded into the repo

* Don't scan deleted files with PHPStan

* Add `dd-` shortcut for PHPStan