Send content to non-JS stream
Because JS isn't enabled, nothing is streamed back (it would be rendered in the browser). Instead, we just send a redirect when the process is complete.

However, if we send a redirect blank, it will be seen as an empty shutdown, and our generic "something went wrong" error page will be shown. So we send a single byte to ensure that the response completes.