DDPB-3028: Fix "accepted" alarm issues (#136)
* Apply large read timeout on all services

We don't really want to put an nginx timeout on at all, since PHP is responsible for handling requests. However:
- We cannot remove the read timeout completely
- Having a five minute timeout gives us a little extra level of protection from e.g. memory leaks

* Don't include report PDF CSS in body

Because this doesn't extend a template, the "head" block is ignored and this is included inline.

The CSS was initially included in the actual head (via review.html.twig), so this is a duplicate. I believe this leads to our SSL_write error.

* Fix error message attachment

The form field was renamed from "file" to "files

* Increase API nginx timeout to 300s

To match frontend

* Revert change to DocumentController

It shouldn't be part of this PR