DDPB-2997: Increase resilience of prof/PA CSV upload (#131)
* Clear entity manager after processing

This clears up some memory

* Stream output during CSV upload

* Stream CSV upload on client side

Use an AJAX request so we can listen to stream and report back to the user

* Provide more detailed stream responses

* Vaguely handle non-AJAX requests

* Manage memory

* Stop hard-coding AuthToken

* Remove accidental scoop from DDPB-2842

* Move EM clearing to controller

So that we don't touch OrgService for now

* Fix PHPStan errors on branch

* Simplify flash bag usage

* Update tests to check stream output

* Output more useful details to stream

* Use flashes and redirects to upload processing

* Ensure stream is read line-by-line

Previously it would cut off the end of lines then fail to process them because they don't start with a command

* Force garbage collection for memory management

* Initialise session

Stops an error when using `addFlash` after outputting content

* Use renderView instead of render

This method doesn't create a Response object, so doesn't output headers

* Terminate stream

Otherwise it will build a response silently and clear your flash bag

* Garbage collect even without stream

* Redesign PA/prof upload form

* Move upload JavaScript into supporting file

* Add general error handling

* Move response handling into OrgService

* Use RestClient in OrgService

This vastly simplifies the query

* Divide OrgService into atomic methods

* Fix PHPStan errors

* Enable debugging on non-dev environments

This is TEMPORARY to identify why uploads are failing on our infrastructure

* Revert "Enable debugging on non-dev environments"

This reverts commit ca24aa24a7d467c69c7eb4a074116c60527480fd.

* Use dev entrypoint for DDPB-2997

More dodgy practices which I'm using to try and debug

* Revert "Use dev entrypoint for DDPB-2997"

This reverts commit 5b863c26aba69becc18f92b8b57f924b47103513.

* Remove warnings from output

No-one reads them, and the sheer volume is causing a memory leak 🙈

* Send content to non-JS stream

Because JS isn't enabled, nothing is streamed back (it would be rendered in the browser). Instead, we just send a redirect when the process is complete.

However, if we send a redirect blank, it will be seen as an empty shutdown, and our generic "something went wrong" error page will be shown. So we send a single byte to ensure that the response completes.

* Remove test case for warnings

We're now not returning those in the stream

* Resolve merge conflicts

* Remove deprecated CSS

* Show a generic error if the upload process did not complete

* Reinstate upload explanation text

* Hide upload explanation when error occurs

* Remove accidental newlines

* Do CSV chunking on client side

This allows the API to just do REST requests, but also doesn't require Redis because we cache results locally and stream to server.

* Remove streaming changes from API

* Restructure code for readability

* Fix PHPStan errors

* Count added items

It's an array of IDs, not a number

* Use 1-based index for progress logging

Otherwise it's 0/150 to 149/150, which is less helpful

* Remove deprecated code

* Use meta refresh rather than HTTP header

Allows us to stream back empty strings to keep the connection alive

* Account for meta refresh in tests