Don't limit memory when warming cache
We normally limit memory to allow other threads on the server to keep processing. In this case, we can guarantee that this is the only process running so it can use as much memory as it likes.

This was consistently an issue when building the dev cache which led to hard-coding as prod. We can now build what cache is required.