DDPB-3025: Enforce PHP alarm and fix critical errors (#146)
* Turn off FPM access log

It's always going to be "app.php" (or "app_dev.php") and litters the logs

* Only log warnings and above

* Fix head image references

The suggested files don't exist any more.

Also updated other header resources to match Design System template: https://design-system.service.gov.uk/styles/page-template/

* Trigger alarm on all critical PHP errors

* Include full error messsage in RestClientException

Provides better information in the logs, and isn't shown to users anyway.

* Implement `renderError` method

Allows controllers to diplay a custom error message to the user

* Soften RuntimeExceptions to HTTP errors

These aren't genuine errors, they're just pages which aren't available

* Fix broken typehint

* Fix controller PHPStan errors

* Fix RestClient PHPStan errors

* Fix PHPStan entity errors

Mostly just issues with PHPDoc typehints

* Install phpstan-mockery, phpstan-phpunit

* Fix unit test namespace

* Fix PHPStan errors in unit test

* Relax rules for an "inaccessible page"

* Remove echo

* Handle RestClientException properly

RestClient methods can throw RestClientException. The controller should catch this and handle it where possible.

* Expect 404 responses, not 500

500 responses indicate a server error, but in this case it's actually just that the report wasn't found

* Use NotFoundException on inaccessible Reports

* Move report error catch into getReport

It should fallback to a 404 in all use cases

* Fix PHPStan errors

* Correct response codes in tests

* Only set active report if one was unsubmitted

* Increase client unit test timeout to 180s

It takes ~40 seconds to start the task, and it can take ~60s to detect it has finished. On a slow day that doesn't leave enough time in 120s to actually run the tests.

* Replace exception with 404

* Don't try and upload files if request empty

* Import correct service path

This will be much clear once we enable autowiring

* Use 404s on MoneyInController

* Use 404s on MoneyOutController

* Fix typehints on MoneyTransactionTrait

* Fix typo

* Fix attachment of error message

* Fic PHPStan errors with ClientController

Amongst them, avoiding a possible RuntimeException

* Warmup the cache in deployment process

Otherwise the cache is initialised as requested, which is slower.

Caching on initialisation can also be the subject of race conditions [1], particularly because AWS immediately hits the service with three identical healthchecks.

[1] https://github.com/twigphp/Twig/issues/1944

* Don't limit memory when warming cache

We normally limit memory to allow other threads on the server to keep processing. In this case, we can guarantee that this is the only process running so it can use as much memory as it likes.

This was consistently an issue when building the dev cache which led to hard-coding as prod. We can now build what cache is required.

* Set environment with SYMFONY_ENV

This is required to ensure commands are run against the correct environment.

I've also used it in app.php for consistency.

* Add exceptions for common report errors

These extend HttpException so will be turned into 403/404, but allow us to add a custom message.

They also make the code easier to read as the exception matches the scenario.

* Use new exceptions in controllers

* Don't assume client is in organisation