DDPB-3071: Fix critical errors in production (#163)
* Extend RestClientException from HttpException

This means the error code from the API is passed to the client: an uncaught 404 from the API will show a 404 to the browser. Same for 403.

When something goes really wrong and the API throws a 500, this will still go to the client and cause an alarm.

* Fix report/ID types

Previous code was failing because it would try and run `getReport(Report)`

* Don't create reports if you have one

The API restricts you to only creating a new report if all your current ones are submitted. But you can never have an unsubmitted report because we always create a new one on submission.

* Don't assume r.unSubmitDate is defined

It's not defined for NDRs.

This error is ignored when debugging is disabled, so doesn't cause a visible error on non-local environments.

* Remove hidden ID field

The checklist ID is derived from the URL. This field messes up replays, because it will set the ID to null and try to create the checklist on resubmission.

* Fix PHPStan errors

* Make `dd-phpstan` check changed files

Without any extra arguments, `dd-phpstan` will now work the same as CI: it will only check the files which have been changed on this branch

* Use correct method for status code

`getCode` will always return 0

* Remove useless try..catch

RestClientException will now behave the same as NotFoundException

* Set RestClientException::code

It's used elsewhere in the application to identify what the error from the API was.

In the future we might like to use `getStatusCode` instead, but this will retain backwards compatibility.

* Revert "Remove useless try..catch"

This reverts commit ada52d1d16dd8fb688219691e4f6c1b92db9acd7.