DDPB-3095: Deleting a submitted document throws critical error (#167)
* Show a custom error rather than throwing

The user is now told what they're doing wrong, and we're not unnecessarily notified

* Use Response::BAD_REQUEST rather than 404

It's a more reasonable status code, and it uses static properties rather than hard-coding a number

* Use HTTP_METHOD_NOT_ALLOWED for error

* Use HTTP_FORBIDDEN for error

After much discussion with Shaun and AlexE we agreed that this makes sense because the document exists and _could_ be deleted (e.g. by an admin archiving) but the current user is not permitted to do so.