DDPB-3056: Ensure migrations are not run on competing nodes (#172)
* Use locked migrate command

This ensures that if the command is run on multiple containers, only one will actually apply the migrations.

* Clean up MigrateLock command

Fixing PHPStan errors for readability

* Increase migration lock to 5 minutes

Some migrations take a long time, such as large data changes, and we don't want another migration instance to spawn halfway through.

This expiry should be irrelevant anyway, since the command cleans up after itself. It's just a catcher, in case the container shuts down half way through migrations.

* Don't start the application unless the migrations are up-to-date

If multiple containers are started simultaneously, one will acquire the lock and perform the migrations. The others won't stabilise because they won't start the app.

Once the migrations are complete, new containers can stabilise.

This ensures the code is never ahead of the database.