DDPB-3120: Hide "Edit deputy details" for prof/PA users (#186)
* Create shared interface for deputy entites

* Show named deputy declaration info where we can

The deputy could be either a NamedDeputy or a User. The declaration page now supports both.

The "edit" link is disabled for NamedDeputys, because any changes should come through the  OPG and CasRec

* Fix PHPStan errors

* Add behat regions

* Add prof and PA tests

To make sure the named deputy's information is shown and not editable

* Add lay test

Move application status save to before we change the address. This means later tests will test the fixture information, which is closer to "no side effects".

* Remove unnecessary declaration

We've literally just been to this page in the last test. And if it were wrong the next step would fail anyway.

* Remove unused parameter

[skip ci]