DDPB-3140: Properly fix pipeline blockages (#195)
* Expect post-NDR report years to be relative

Rather than hardcoding the expected report years after an NDR is submitted, expect them to be relative to the current year.

* Ensure macro is imported for checklist

* Remove "expected-date" from checklist

It always has the "hidden" class so doesn't provide any user value.

It's also inaccurate, since it always refers to the current year rather than the year the report should be submitted.

* Stop trying to force HTTP 1.0

The API only supports HTTP 1.1. By trying to force 1.0, Guzzle was handling the request differently and failing requests with the latest version of curl.