Move healthcheck to file scanner server
This looks a bit messy, but does exactly what we need.

The file scanner REST container now won't start until the server is *explicitly* healthy. The overall task waits for both containers to start, so we can now ensure that no traffic will be sent to the task until both containers are ready.