DDPB-3126: App briefly crashes as ClamAV updates (#199)
* Tell Clamd to only self-check

Freshclam will update the database in the background, but Clamd will be responsible for reloading it into the application.

This should stop conflicts from two reload procedures. It also generally simplifies the configuration.

* Try using notify rather than SelfCheck

I've confirmed that SelfCheck causes a ~30 second outage

* Disable self-updating in ClamAV

* Design redeployer Go script

* Build redeployer in CI

* Create a shared lambda for redeploying

This can be used to redeploy any account service on demand

* Create schedule to redeploy file scanner

It will be redeployed once per day, causing the virus definitions to be updated

* [test] Apply shared development in PR

Want to check this works, so doing it in the PR rather than just on master

* Fix name of redeployer build file

* Fix name of lambda handler

* Tidy up code

* Remove test steps

* Move redeployer code to shared repo

It's a more natural home, and there's no need for it to overlap with other Go code

* Use new redeployer filepath in CI

* Add apply shared-development again

To check the file move worked

* Fix Lambda function name

* Add healthcheck to scan container

This ensures ECS won't treat the container as healthy until it can accept HTTP requests. During that time, the old container will remain active.

* Fix syntac of healthcheck

* Remove test steps

* Use a  healthcheck that sets status code

The old helathcheck just returned 200 or 5xx, which would always be treated as "healthy"

This now returns an error status code (22) if there was a problem, which should trigger "unhealthy"

* Remove healthcheck

This is used for turning a service off when it's not healthy. What I want is to keep it in PENDING until it's ready.

* Move healthcheck to file scanner server

This looks a bit messy, but does exactly what we need.

The file scanner REST container now won't start until the server is *explicitly* healthy. The overall task waits for both containers to start, so we can now ensure that no traffic will be sent to the task until both containers are ready.

* Fix README errors

* Ensure Lambda can write to logs

* Build Go binary for correct architecture