DDPB-3168: Don't create new report when NDR submitted (#206)
* Fix test

This only worked because the report submission ID and report ID happened to be the same (1). We now get the report submission properly.

* Add test for submitting NDRs with existing report

In this case, the existing report should be given the assets/bank accounts and no new report should be created

* Better handle NDR submissions

If a report exists, copy resources across. Otherwise, create a new one.