Disable npm audit
Node Advisory 1488 [1] is blocking our build and is currently unavoidable.

It's a dependency of Webpack, but they would need need to push a breaking change [2]. It's also a dependency of sass-lint [3], but that hasn't been updated in over two years.

We could remove sass-lint, but we're reliant on Webpack. The vulnerability would cause build issues locally and on CircleCI, but is very unlikely to occur. It wouldn't affect service availability or allow data exfiltration.

[1] https://www.npmjs.com/advisories/1488
[2] https://github.com/webpack/webpack/issues/10516
[3] https://github.com/sasstools/sass-lint