Add check when registering lay deputies
When a lay deputy tries to register, we should check that that client doesn't already have a court order attached to a user or organisation.

Right now this is duplcaiting line 62, but in the future it will be necessary as client's won't have deputies assigned through the previous method.

There's no harm to adding it now since no-one will be assigned via court order who wouldn't be caught on line 62