DDPB-3315: Illogical heading structure
Heading that are of the same semantic level should use the same HTML
heading level tag. For example all <h2> but when a sub-heading is added
it should increment the level by one <h3>. This ticket does that while
also removing <h4> as they should rarely be needed.

Fixes DDPB-3315