DDPB-3667: refactor the client AbstractController to prepare for symfony upgrade
In the client section of the app we were using a base controller called
AbstractController that the majority of our controllers extended from.
This PR attempts to clean that up by removing the functions in the
controller to there own services and injecting them into the classes
that require them.

Fixes DDPB-3667