DDPB-3786: Handle larger text in the feedback form and prevent minor spam
We had a limit of 1200 characters at the DB level for the feedback form
but did not apply the limit to the frontend. We now just increased the
DB character count to a lot more by changing the field to a TEXT field.

Also we added a honeypot method to the feedback form to try and stop
random (non-targeted) spam.

Issue #DDPB-3786