[DDPB-4353] add created_at and updated_at to entities to provide context
Added a created_at and updated_at column to the reports, clients, users,
and named_deputies tables. Created a timestamps trait to handle most of
this and added the ORM\HasLifecycleCallbacks() to each entity that will
use this call back for either presisting this data or when the data is
updated.

Also added a migration file to add these columns to the needed tables